requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pyjwt==2.8.0
python-decouple==3.8
email_validator==2.1.1
